Extract sample-wise coefficients (alpha_i) from a fastsvm model

This method returns the vector of coefficients coef_ estimated by sksurv.svm.FastKernelSurvivalSVM. Each coefficient corresponds to the weight assigned to an individual training sample in the kernel-induced decision function.

# S3 method for class 'fastsvm'
coef(object, ...)

Arguments

object

An object of class "fastsvm" returned by fast_kernel_surv_svm_fit().

...

Additional arguments (unused; included for S3 compatibility).

Value

A numeric vector of length n, containing the sample-wise coefficients alpha_i.

Details

These coefficients play the same role as support vector weights in classical SVMs: samples with non-zero coefficients (within a tolerance) can be interpreted as "support-like" vectors.

Examples

if (reticulate::py_module_available("sksurv")) {
  set.seed(1)
  df <- data.frame(
    time = rexp(50, 0.1),
    status = rbinom(50, 1, 0.7),
    x1 = rnorm(50),
    x2 = rnorm(50)
  )

  fit <- fast_kernel_surv_svm_fit(
    data = df,
    time_col = "time",
    delta_col = "status",
    kernel = "rbf"
  )

  coef(fit)  # extract coefficients
}
#>  [1]  1.798843e-01  9.738373e-01 -1.006409e+00 -8.896473e-01 -5.484181e-01
#>  [6]  1.014643e+00  4.106031e-01 -1.765767e-02 -1.386360e-01  1.501420e-01
#> [11]  8.192291e-01 -6.035224e-01  4.979419e-02  5.926589e-01  5.501324e-01
#> [16]  8.581404e-01  1.037989e+00  2.843310e-01 -5.542113e-01 -4.495055e-01
#> [21]  1.359887e-02 -4.069243e-01  1.635826e-01 -4.332815e-02 -6.341219e-05
#> [26] -2.895365e-01  2.477798e-01  1.802034e+00  1.158698e+00  3.217526e-01
#> [31]  8.372927e-01 -2.304474e+00 -7.067208e-01  2.226941e-01  2.692737e-02
#> [36]  1.698688e-01 -4.842756e-01  5.294004e-01  2.627820e-01 -1.022173e+00
#> [41]  5.670927e-01  9.150460e-01  7.897894e-01 -7.342176e-03 -4.869664e-02
#> [46] -2.904666e-01  1.620772e-01  9.066525e-01  2.057757e-01  1.246496e+00